+++ packages/jsii-rosetta/lib/translate.ts
@@ -67,7 +67,7 @@ export class Translator {
       location: snip.location,
       didCompile: translator.didSuccessfullyCompile,
       fqnsReferenced: translator.fqnsReferenced(),
-      fullSource: snip.completeSource,
+      fullSource: completeSource(snip),
       syntaxKindCounter: translator.syntaxKindCounter(),
     });
   }
@@ -152,6 +152,7 @@ export class SnippetTranslator {
   public readonly compileDiagnostics: ts.Diagnostic[] = [];
   private readonly visibleSpans: Spans;
   private readonly compilation!: CompilationResult;
+  private readonly tryCompile: boolean;
 
   public constructor(snippet: TypeScriptSnippet, private readonly options: SnippetTranslatorOptions = {}) {
     const compiler = options.compiler ?? new TypeScriptCompiler();
@@ -170,7 +171,9 @@ export class SnippetTranslator {
     this.visibleSpans = Spans.visibleSpansFromSource(source);
 
     // This makes it about 5x slower, so only do it on demand
-    if (options.includeCompilerDiagnostics || snippet.strict) {
+    // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
+    this.tryCompile = (options.includeCompilerDiagnostics || snippet.strict) ?? false;
+    if (this.tryCompile) {
       const program = this.compilation.program;
       const diagnostics = [
         ...neverThrowing(program.getGlobalDiagnostics)(),
@@ -196,7 +199,7 @@ export class SnippetTranslator {
         try {
           return call(...args);
         } catch (err) {
-          const isExpectedTypescriptError = err.message.includes('Error: Debug Failure');
+          const isExpectedTypescriptError = err.message.includes('Debug Failure');
 
           if (!isExpectedTypescriptError) {
             console.error(`Failed to execute ${call.name}: ${err}`);
@@ -208,8 +211,11 @@ export class SnippetTranslator {
     }
   }
 
+  /**
+   * Returns a boolean if compilation was attempted, and undefined if it was not.
+   */
   public get didSuccessfullyCompile() {
-    return this.compileDiagnostics.length === 0;
+    return this.tryCompile ? this.compileDiagnostics.length === 0 : undefined;
   }
 
   public renderUsing(visitor: AstHandler<any>) {